+++ tests/UtilsTests/test_matrices.c
@@ -136,7 +136,7 @@ Test(matrices, test_scalar_multiply) {
     cr_assert(matrice_equals(m1, expected), "Matrices are not equal");
 }
 
-double multiply_by_two(double x) { return x * 2; }
+float multiply_by_two(float x) { return x * 2; }
 
 Test(matrices, test_map) {
     setup();
@@ -196,7 +196,7 @@ Test(matrices, test_max) {
     int i;
     int j;
 
-    double *max = matrice_max(m1, &i, &j);
+    float *max = matrice_max(m1, &i, &j);
 
     cr_assert(i == 1 && j == 1, "Max is not in the right place");
     cr_assert(*max == 4, "Max is not equal to 4");
@@ -227,7 +227,7 @@ Test(matrices, test_max_no_index) {
     // value, so we should be able to pass NULL
     // as the index parameters
 
-    double *max = matrice_max(m1, NULL, NULL);
+    float *max = matrice_max(m1, NULL, NULL);
 
     cr_assert(*max == 4, "Max is not equal to 4");
 }
@@ -235,7 +235,7 @@ Test(matrices, test_max_no_index) {
 Test(matrices, test_sum) {
     setup();
 
-    double sum = matrice_sum(m1);
+    float sum = matrice_sum(m1);
 
     cr_assert(sum == 10, "Sum is not equal to 10");
 }
@@ -243,7 +243,7 @@ Test(matrices, test_sum) {
 Test(matrices, test_mean){
     setup();
 
-    double mean = matrice_mean(m1);
+    float mean = matrice_mean(m1);
 
     cr_assert(mean == 2.5, "Mean is not equal to 2.5");
 }
@@ -251,7 +251,7 @@ Test(matrices, test_mean){
 Test(matrices, test_std){
     setup();
 
-    double std = matrice_std(m1);
+    float std = matrice_std(m1);
 
     cr_assert(std == 1.118033988749895, "Std is not equal to 1.118033988749895");
 }
